use LONGLONG instead of long long for consistency
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@5195 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/parser/MP4Splitter/Ap4AsyncReaderStream.cpp
@@ -54,7 +54,7 @@ AP4_Result AP4_AsyncReaderStream::Read(void* buffer, AP4_Size bytesToRead, AP4_S
 {
     __int64 bytesAvail = m_pFile->GetRemaining();
 
-    if (bytesAvail < (long long)bytesToRead) {
+    if (bytesAvail < (LONGLONG)bytesToRead) {
         if (bytesRead) {
             *bytesRead = bytesAvail;
         }
